% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima}
\alias{arima}
\title{Constructor for arima object}
\usage{
arima(phi, theta, delta)
}
\arguments{
\item{phi, theta}{numeric vectors of length \eqn{\ge 0} giving AR
    and MA parameters.}

\item{delta}{vector of differencing coefficients, so an ARMA model is fitted
to \code{y[t] - Delta[1]*y[t-1] - ...}.}
}
\description{
Constructor for arima object
}
