% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_arima.R
\name{extract_arima}
\alias{extract_arima}
\alias{extract_arima._Arima_fit_impl}
\alias{extract_arima.Arima}
\title{Extract arima object}
\usage{
extract_arima(object, ...)

\method{extract_arima}{`_Arima_fit_impl`}(object, ...)

\method{extract_arima}{Arima}(object, ...)
}
\arguments{
\item{object}{Object of class "Arima" or "_Arima_fit_impl".}

\item{...}{not used.}
}
\value{
"arima" object.
}
\description{
Extract underlying \code{arima} object from fitted ARIMA model.
}
