% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{fitted_resid.arima}
\alias{fitted_resid.arima}
\title{Fitted residuals for ARIMA}
\usage{
\method{fitted_resid}{arima}(object, new_data, ...)
}
\arguments{
\item{object}{a registered model. Currently supports objects of class \code{arima}
and \code{garch}.}

\item{new_data}{a \link[tsibble:tsibble-package]{tsibble} of univariate
time series on which residuals are to be calculated.}

\item{...}{not used}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} of fitted residuals of class
\code{fitted_resid_ts}.
}
\description{
Fitted residuals for ARIMA
}
