% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simults.R
\name{fitted_resid}
\alias{fitted_resid}
\alias{fitted_resid.default}
\title{Fitted residuals from registered models}
\usage{
fitted_resid(object, ...)

\method{fitted_resid}{default}(object, ...)
}
\arguments{
\item{object}{a registered model. Currently supports objects of class \code{arima}
and \code{garch}.}

\item{...}{not used}
}
\description{
Fitted residuals from registered models
}
