% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{inn2ts.arima}
\alias{inn2ts.arima}
\title{Apply ARIMA recursions}
\usage{
\method{inn2ts}{arima}(object, inn, ...)
}
\arguments{
\item{object}{an \code{arima} object}

\item{inn}{vector of innovations.}

\item{...}{not used.}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} of class "arima_ts". Column
"ts" corresponds to the simulated time series and column "date" has
dummy dates for each observation.
}
\description{
This implementation is based on \code{stats::arima.sim}.
}
\examples{
# simulate from an AR(1) model
mdl <- make_arima(phi = 0.3, theta = 0, delta = 0)
inn2ts(mdl, inn = rnorm(200))
}
