% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{inn2ts.garch}
\alias{inn2ts.garch}
\title{Custom simulation of univariate GARCH(1,1) model}
\usage{
\method{inn2ts}{garch}(object, inn, ...)
}
\arguments{
\item{object}{A \code{garch} object.}

\item{inn}{vector of innovations.}

\item{...}{not used.}
}
\value{
A \link[tsibble:tsibble-package]{tsibble} of class "garch_ts". Column
"ts" corresponds to the simulated time series and column
"date" has dummy dates for each observation.
}
\description{
Simulates a univariate time series according to \eqn{X_t  = \sigma_t Z_t},
where \eqn{\sigma_t} follows the recursion
\deqn{\sigma_{t|t-1} = \sum_{j=1}^p \beta_j \sigma_{t-j|t-j-1}+ \alpha_0+\sum_{j=1}^q \alpha_j X_{t-i}^2.}
}
\details{
We fix \eqn{p=q=1} for the GARCH model. The implementation is based on \code{fGarch::garchSim}.
}
\examples{
# simulate from a GARCH(1, 1) model
mdl <- make_garch(omega = 0.5, alpha = 0.1, beta = 0.8)
inn2ts(
     mdl,
     inn = rnorm(200)
)
}
