% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{inn2ts}
\alias{inn2ts}
\title{Simulate observations}
\usage{
inn2ts(object, inn, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} and \code{garch}.}

\item{inn}{vector of innovations.}

\item{...}{not used.}
}
\description{
This generic is used to obtain observations from a time series model given innovations.
}
