% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcits-package.R
\docType{package}
\name{rcits-package}
\alias{rcits}
\alias{rcits-package}
\title{rcits: Recursions for Causal Invertible Time Series}
\description{
Given a causal and invertible time series, simulates time series observations given innovations or vice-versa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonfernandes/rcits}
  \item Report bugs at \url{https://github.com/leonfernandes/rcits/issues}
}

}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
