% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simts-package.R
\docType{package}
\name{simts-package}
\alias{simts}
\alias{simts-package}
\title{simts: Simulate Time Series}
\description{
Simulates data from various time series models. Also includes helper functions to specify as well as extract underlying model parameters from fitted objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonfernandes/simts}
  \item Report bugs at \url{https://github.com/leonfernandes/simts/issues}
}

}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
