% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simts.R
\name{simts}
\alias{simts}
\title{Simulate from registered models}
\usage{
simts(object, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} or fitted using \code{stats::arima} and \code{modeltime::arima_reg}.}

\item{...}{not used.}
}
\description{
Simulate from registered models
}
