% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simults-package.R
\docType{package}
\name{simults-package}
\alias{simults-package}
\alias{_PACKAGE}
\alias{simts-package}
\title{simults: Simulate Time Series}
\description{
Simulates data from arima and garch models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonfernandes/simults}
  \item Report bugs at \url{https://github.com/leonfernandes/simults/issues}
}

}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
