% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{simults.arima}
\alias{simults.arima}
\title{Simulate from arima models}
\usage{
\method{simults}{arima}(object, nsim, innov, ...)
}
\arguments{
\item{object}{a fitted arima model}

\item{nsim}{positive integer. Length of time series to be generated.}

\item{innov}{vector of innovations.}

\item{...}{not used.}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} of class "arima_ts". Column
"value" corresponds to the simulated time series and column "date" has
dummy dates for each observation.
}
\description{
The innovations \code{innov} are assumed to include the burn-in phase. Only the
last \code{nsim} simulations are returned.
The implementation is based on \code{stats::arima.sim}.
}
\examples{
# simulate from an AR(1) model
mdl <- make_arima(phi = 0.3, theta = 0, delta = 0)
simults(mdl, nsim = 100, innov = rnorm(200))
}
