% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simults.R
\name{simults}
\alias{simults}
\alias{simults.default}
\title{Simulate from registered models}
\usage{
simults(object, ...)

\method{simults}{default}(object, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} and \code{garch}.}

\item{...}{not used.}
}
\description{
Simulate from registered models
}
