% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{simults.garch}
\alias{simults.garch}
\title{Custom simulation of univariate GARCH model}
\usage{
\method{simults}{garch}(object, nsim, innov, init_garch, init_sigma2, ...)
}
\arguments{
\item{object}{A \code{garch} object.}

\item{nsim}{positive integer. Number of simulations.}

\item{innov}{numeric vector of innovations.}

\item{init_garch}{numeric vector of starting values for garch process
used only for burn-in period.}

\item{init_sigma2}{numeric vector of starting values for variance process
used only for burn-in period.}

\item{...}{Not used.}
}
\value{
A \link[tsibble:tsibble-package]{tsibble} of class "garch_ts". Column
"value" corresponds to the simulated time series, "sigma2" contains the
simulated variance time series and column "date" has dummy dates per
observation.
}
\description{
Simulates a univariate time series according to \eqn{X_t  = \sigma_t Z_t},
where \eqn{\sigma_t} follows the recusrion
\deqn{\sigma_{t|t-1} = \sum_{j=1}^p \beta_j \sigma_{t-j|t-j-1}+ \alpha_0+\sum_{j=1}^q \alpha_j X_{t-i}^2.}
It is assumed that sufficient terms for burn-in period are provided in
\code{innov}, \code{init_garch} and \code{init_sigma2}. The number of terms used for burn-in
is \eqn{m}, where \eqn{m} is the minimum lengths oamong \code{init_garch} and
\code{init_sigma2}. In this case \code{innov} is assumed to have at least \code{nsim} +
\eqn{m}. Note that only the minimal required number of terms, counted from the
tail are used for the simulation.
The implementation is based on \code{fGarch::garchSim}.
}
\examples{
# simulate from a GARCH(1, 1) model
mdl <- smpspltools::make_garch(omega = 0.5, alpha = 0.1, beta = 0.8)
simults(
     mdl,
     nsim = 100,
     innov = rnorm(200),
     init_garch = c(0, 0),
     init_sigma2 = c(1, 1)
)
}
