% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_model.R
\name{translate_model}
\alias{translate_model}
\alias{translate_model.default}
\alias{translate_model.arima}
\alias{translate_model.Arima}
\alias{translate_model.model_fit}
\alias{translate_model.workflow}
\alias{translate_model.Arima_fit_impl}
\alias{translate_model.mdl_ts}
\alias{translate_model.ARIMA}
\alias{translate_model.AR}
\alias{translate_model.garch}
\alias{translate_model.fGARCH}
\title{Extracts available models}
\usage{
translate_model(object, ...)

\method{translate_model}{default}(object, ...)

\method{translate_model}{arima}(object, ...)

\method{translate_model}{Arima}(object, ...)

\method{translate_model}{model_fit}(object, ...)

\method{translate_model}{workflow}(object, ...)

\method{translate_model}{Arima_fit_impl}(object, ...)

\method{translate_model}{mdl_ts}(object, ...)

\method{translate_model}{ARIMA}(object, ...)

\method{translate_model}{AR}(object, ...)

\method{translate_model}{garch}(object, ...)

\method{translate_model}{fGARCH}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{unused.}
}
\description{
Generic function to extract and return objects that have been registered with
\code{rcits} package. Currently only \code{arima} and \code{garch} objects are registered.
This function translates functions of class \code{Arima}, \code{Arima_fit_impl}, \code{AR},
\code{ARIMA} and \code{fGARCH} to the corresponding registered methods. This function
also extracts the fitted models from \code{model_fit}, \code{workflow} and \code{mdl_ts}
objects.
}
