% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{ts2inn.garch}
\alias{ts2inn.garch}
\title{Fitted residuals for GARCH(1,1)}
\usage{
\method{ts2inn}{garch}(object, ts, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} and \code{garch}.}

\item{ts}{a \link[tsibble:tsibble-package]{tsibble} of univariate time series
observations.}

\item{...}{not used}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} where column \code{inn}
contains the calculated innovations.
}
\description{
Fitted residuals for GARCH(1,1)
}
