% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ts2inn}
\alias{ts2inn}
\title{Fitted residuals from registered models}
\usage{
ts2inn(object, ts, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} and \code{garch}.}

\item{ts}{a \link[tsibble:tsibble-package]{tsibble} of univariate
time series observation}

\item{...}{not used}
}
\description{
This generic is used to obtain innovations from a time series model given observations.
}
