% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simts.R
\name{simts}
\alias{simts}
\alias{simts.default}
\title{Simulate from registered models}
\usage{
simts(object, ...)

\method{simts}{default}(object, ...)
}
\arguments{
\item{object}{model from which to simulate. Currently supports objects of
class \code{arima} and \code{garch}.}

\item{...}{not used.}
}
\description{
Simulate from registered models
}
