% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{simts.arima}
\alias{simts.arima}
\alias{simts._Arima_fit_impl}
\alias{simts.Arima}
\title{Simulate from arima models}
\usage{
\method{simts}{arima}(object, nsim, innov, ...)

\method{simts}{`_Arima_fit_impl`}(object, nsim, innov, ...)

\method{simts}{Arima}(object, nsim, innov, ...)
}
\arguments{
\item{object}{either \code{arima}, \code{Arima} or \verb{_Arima_fit_impl} object.}

\item{nsim}{positive integer. Length of time series to be generated.}

\item{innov}{vector of innovations.}

\item{...}{not used.}
}
\value{
a \link[tibble:tibble-package]{tibble} of class "arima_tbl". Column "x"
corresponds to the simulated time series and column "date" has dummy
dates for each observation.
}
\description{
The innovations \code{innov} are assumed to include the burn-in phase. Only the
last \code{nsim} simulations are returned.
The implementation is based on \code{stats::arima.sim}.
}
